Saving changes for merge

+++ server/api/products.js
@@ -24,28 +24,31 @@ router.get('/:productId', (req, res, next) => {
 })
 
 router.post('/addProduct', (req, res, next) => {
-  Product.create(req.body)
-    .then(createdProduct => res.status(201).json(createdProduct))
-    .catch(next)
+  if (req.user.isAdmin)
+    Product.create(req.body)
+      .then(createdProduct => res.status(201).json(createdProduct))
+      .catch(next)
 })
 
 router.delete('/:productId', (req, res, next) => {
-  Product.destroy({
+  if (req.user.isAdmin)
+    Product.destroy({
       where: {
         id: +req.params.productId
       }
     })
-    .then(() => res.sendStatus(204))
-    .catch(next)
+      .then(() => res.sendStatus(204))
+      .catch(next)
 })
 
 router.put('/:productId', (req, res, next) => {
-  Product.update(req.body, {
+  if (req.user.isAdmin)
+    Product.update(req.body, {
       where: {
         id: +req.params.productId
       },
       returning: true
     })
-    .then(([rows, [product]]) => res.json(product))
-    .catch(next)
+      .then(([rows, [product]]) => res.json(product))
+      .catch(next)
 })
